fix up context menus
